Added check if config parameters are up to date.
CmiSchedule:
Added methods for comparing the xml loaded equipment config data with
web parsed settings and options. If these are not compatible, the
equipment will be locked for reservations.

ScheduleManager:
Checks if equipment is locked after web loading and parsing has
finished, and blocks attemts to make reservations if necessary.

Configuration:
Can now be instantiated without a valid reference to an equipment. 

Equipment:
Added lock(); and isLocked(); methods.