Changes to Configuration
- Setting.findOption() now also finds Option object by value, in
addition
to name and title. 

- New method Configuration.findSetting() finds Setting objects by id and
name.

- Added class Values that incorporates an "instance" of a configuration,
storing value name pairs for settings and options, without the complete
structure of the configuration tree. This is experimental...

- Added Copy constructor, creating a deep copy of all nested classes,
except the Option classes which is copied by reference. 

ACHTUNG! This functionality is INCOMPLETE! While the Configuration Node
Tree is copied, the Settings list is not. I still have to find a smart
way to do that.

- Added clone methods for Node, Setting, Group and Option. 

Changes to WebLoadedConfiguration

- Added static method parseHoverLine, returning Configuration.Values
from
the onmouseover attribute of the td elements in the schedule web table.