Corrections to CmiAccount CmiServerConnection etc.
- CmiAccount:
 added shorthand methods for SshTunnel control

- CmiServerConnections: 
  fixed a few errors. 
 
- CmiHttpRequest:
  fixed errors in how Cookies and data strings were composed. 
  
- CmiSshTunnel:
  Checks if tunnel is active by checking if the session object is not
null and is connected. Note that this only works if the connection is
properly closed. In case where the CmiSshTunnel object is discarded
without closing the tunnel using close(), the tunnel will remain open
beyond the lifetime of the object. A new CmiSshTunnel object is
created it will erroneously report a closed tunnel while in reality
it is still open. Therefore only one SshTunnel object should exist at
any time. 
  Actually this means I should make it a singleton!
Implemented finalize() to close the tunnel when the object gets garbage 
collected. But this apparently is not guaranteed to be called.