changes in CmiServerConnection:
getMainPage:
	Overloaded to accept dateIndex parameter (int)
	Overloaded to accept Configuration.Values parameter
	... and any combination of these and machId

getAllReservationsPage
	Overloaded to accept LocalDate and timeStamp (String) parameters
	
createHttpConnection
	setRequestProperty "Connection:close" for new connections. 

Adressed (maybe not excatly "fixed") a bug where connecting to the
server was impossible after a previous connection had been made
(EOFException). This bug occurred as an error when a new connection was
made and the attempt was made to configure the properties of the new
connection, like setRequestMethod or setDoOutput. The error reported the
connection was already open although, the Connection object was just
newly instantiated. The reason for this behavior is connection pooling.
The framework keeps the connection alive behind the scenes and reuses
it. 

The line following line in createHttpConnection helps prevent this
problem: 
this.connection.setRequestProperty("Connection", "close");

In addition the connection always has to be explicitely closed by the
client (via the close() methods of the inputStream or the reader). 

See also: 
http://stackoverflow.com/questions/19258518/android-httpurlconnection-eofexception