CmiAccount: getActive / setActive
CmiAccount in it's current implementation is some kind of a half-baked
singleton. It has a static "instance" variable that holds *the* instance
of CmiAccount and an instance() method that returns it. 

However, the class can still be freely instantiated. 

So we now have the static getActive method that still returns *the*
instance, same as instance(). We also have a non-static setActive()
method that promotes the CmiAccount object to be *the* instance.